requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
cowsay==6.1
Flask==3.0.3
Flask-WTF==1.2.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
